Divide logos into groups
