--- model/logo.js
@@ -1,76 +1,0 @@
-var logoList = [
-  {
-    title: "CodeHub",
-    url: "http://www.codehub.org.uk/",
-    src: "/img/logo/codehub.png",
-    alt: "CodeHub is a place to code, share, and learn from one another. If you are developing for the web and wish to exchange knowledge and skills, come and join us. We are largely based within a 25 mile radius of Bristol/Bath, but anyone welcome."
-  },
-  {
-    title: "Meetup",
-    url: "https://www.meetup.com/CodeHub-Bristol/",
-    src: "/img/logo/meetup.png",
-    alt: "Meetup!"
-  },
-  {
-    title: "Bristol JS",
-    url: "https://www.meetup.com/BristolJS/",
-    src: "/img/logo/bristol-js.png",
-    alt: "Bristol JS"
-  },
-  {
-    title: "Beetroot Cafe",
-    url: "http://www.beatrootcafe.com/",
-    src: "/img/logo/beetroot.png",
-    alt: "Meetup!"
-  },
-  {
-    title: "Gitter",
-    url: "https://gitter.im/CodeHubOrg/discussions/",
-    src: "/img/logo/gitter.png",
-    alt: "Gitter"
-  },
-  {
-    title: "Github",
-    url: "https://github.com/CodeHubOrg",
-    src: "/img/logo/github.png",
-    alt: "Github"
-  },
-  {
-    title: "JavaScript",
-    url: "https://developer.mozilla.org/en-US/docs/Web/JavaScript/",
-    src: "/img/logo/js.png",
-    alt: "JavaScript"
-  },
-  {
-    title: "Node JS",
-    url: "https://nodejs.org/",
-    src: "/img/logo/nodejs.png",
-    alt: "Node JS"
-  },
-  {
-    title: "NPM",
-    url: "https://www.npmjs.com/",
-    src: "/img/logo/npm.png",
-    alt: "NPM"
-  },
-  {
-    title: "React",
-    url: "https://facebook.github.io/react/",
-    src: "/img/logo/react.png",
-    alt: "React"
-  },
-  {
-    title: "Ubuntu",
-    url: "https://www.ubuntu.com/",
-    src: "/img/logo/ubuntu.png",
-    alt: "Ubuntu"
-  },
-  {
-    title: "The God Father of JavaScript - Douglas Crockford",
-    url: "http://bdcampbell.net/javascript/book/javascript_the_good_parts.pdf",
-    src: "/img/logo/douglas-crockford.png",
-    alt: "The God Father of JavaScript - Douglas Crockford"
-  }
-];
-
-module.exports = logoList
