+++ model/meetup.js
@@ -1,0 +1,16 @@
+var logoList = [
+  {
+    title: "CodeHub",
+    url: "https://www.meetup.com/CodeHub-Bristol/",
+    src: "/img/logo/codehub.png",
+    alt: "CodeHub is a place to code, share, and learn from one another. If you are developing for the web and wish to exchange knowledge and skills, come and join us. We are largely based within a 25 mile radius of Bristol/Bath, but anyone welcome."
+  },
+  {
+    title: "Bristol JS",
+    url: "https://www.meetup.com/BristolJS/",
+    src: "/img/logo/bristol-js.png",
+    alt: "Bristol JS"
+  }
+];
+
+module.exports = logoList
