+++ model/tech.js
@@ -1,0 +1,40 @@
+var logoList = [
+  {
+    title: "JavaScript",
+    url: "https://developer.mozilla.org/en-US/docs/Web/JavaScript/",
+    src: "/img/logo/js.png",
+    alt: "JavaScript"
+  },
+  {
+    title: "Node JS",
+    url: "https://nodejs.org/",
+    src: "/img/logo/nodejs.png",
+    alt: "Node JS"
+  },
+  {
+    title: "NPM",
+    url: "https://www.npmjs.com/",
+    src: "/img/logo/npm.png",
+    alt: "NPM"
+  },
+  {
+    title: "React",
+    url: "https://facebook.github.io/react/",
+    src: "/img/logo/react.png",
+    alt: "React"
+  },
+  {
+    title: "Github",
+    url: "https://github.com/CodeHubOrg",
+    src: "/img/logo/github.png",
+    alt: "Github"
+  },
+  {
+    title: "Ubuntu",
+    url: "https://www.ubuntu.com/",
+    src: "/img/logo/ubuntu.png",
+    alt: "Ubuntu"
+  }
+];
+
+module.exports = logoList
